Creates csv_data class to abstract the csv data loading
